requirements.txt
asgiref==3.2.5
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
idna==2.9
oauthlib==3.1.0
pkg-resources==0.0.0
psycopg2==2.8.4
PyOpenGraph==0.2
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
webpreview==1.6.0
whitenoise==5.0.1
